Add files via upload
The aim of Traveling Salesman Problem is to help a salesman to find the shortest route for a round trip by visiting every city only once. As this problem is a impossible to solve in polynomial time, it is classified as NP (Non-Polynomial problem). The problem in this project is identified by applying Travelling Salesman Problem. It can be used as an example for a shipping company to create it's transport network with the lowest cost and can also be used for internet network traffic protocol problems. It is aimed to solve a problem with understanding the data structures and models, using graph structures and algorithm logic by the project doer.